oog/qr_reader/requirements.txt
pillow
pyqrcode
